+++ Arbiter.App/Models/Settings/DebugSettings.cs
@@ -6,7 +6,6 @@ public class DebugSettings : ICloneable
 {
     public bool ShowDialogId { get; set; }
     public bool ShowPursuitId { get; set; }
-    public bool ShowDialogItemQuantity { get; set; }
     public bool ShowEquipmentDurability { get; set; }
     public bool ShowNpcId { get; set; }
     public bool ShowMonsterId { get; set; }
@@ -27,7 +26,6 @@ public class DebugSettings : ICloneable
     {
         ShowDialogId = ShowDialogId,
         ShowPursuitId = ShowPursuitId,
-        ShowDialogItemQuantity = ShowDialogItemQuantity,
         ShowEquipmentDurability = ShowEquipmentDurability,
         ShowNpcId = ShowNpcId,
         ShowMonsterId = ShowMonsterId,